ingestion/usda_data_ingestor.py
# This script is the first step in our new MLOps architecture.
# It is responsible for ingesting fundamental agricultural data from the
# free USDA Quick Stats API and preparing it for our BigQuery data warehouse.

import os
import requests
from google.cloud import bigquery

# It's a best practice to manage API keys via environment variables or a secret manager.
USDA_API_KEY = os.environ.get("USDA_API_KEY")
PROJECT_ID = os.environ.get("GCP_PROJECT")
DATASET_ID = "commodity_data"
TABLE_ID = "usda_fundamentals"

def fetch_usda_data():
    """Fetches key soybean data from the USDA Quick Stats API."""
    if not USDA_API_KEY:
        raise ValueError("USDA_API_KEY environment variable not set.")
    
    # Example query: US Soybean stocks, production, and acreage
    # This will be expanded to cover all the fundamental factors you listed.
    params = {
        'key': USDA_API_key,
        'commodity_desc': 'SOYBEANS',
        'statisticcat_desc': ['STOCKS', 'PRODUCTION', 'ACREAGE'],
        'agg_level_desc': 'NATIONAL',
        'year__GE': '2000', # Get data from the year 2000 onwards
        'format': 'JSON'
    }
    
    url = "https://quickstats.nass.usda.gov/api/get_params/"
    print("Fetching data from USDA Quick Stats API...")
    # In a real implementation, we would make a POST request with these params
    # response = requests.post(url, json=params)
    # response.raise_for_status()
    # print("Data fetched successfully.")
    # return response.json()
    
    # For now, return a placeholder
    return {"data": "placeholder"}

def load_data_to_bigquery(data):
    """Loads the fetched data into a BigQuery table."""
    if not PROJECT_ID:
        raise ValueError("GCP_PROJECT environment variable not set.")

    client = bigquery.Client(project=PROJECT_ID)
    table_ref = client.dataset(DATASET_ID).table(TABLE_ID)

    # In a real implementation, we would define the schema and load the data.
    print(f"Would load data into BigQuery table: {PROJECT_ID}.{DATASET_ID}.{TABLE_ID}")
    # job = client.load_table_from_json(data, table_ref, job_config=job_config)
    # job.result() # Wait for the job to complete
    # print(f"Loaded {job.output_rows} rows into {table_ref.path}")

if __name__ == "__main__":
    # This demonstrates the new workflow.
    # We will need to set up the USDA_API_KEY in our environment to run this.
    print("Starting USDA data ingestion script...")
    # usda_data = fetch_usda_data()
    # load_data_to_bigquery(usda_data)
    print("Script placeholder executed. Next step: acquire API key and define schema.")
